remove obsolete GrapheneOS pie branch
This branch no longer exists. It had not been actively developed since
August 2019 when GrapheneOS migrated to Android 10 and was obsolete when
it was added to the list here. The final state of the pie branch is
available through the PQ3B.190801.002.2019.08.25.15 tag.

The branches in GrapheneOS repositories forked from AOSP are development
workspaces with rebased patch sets. The long-term history is available
via the tags released at least once per month, not the branches. It's
maintained this way to keep it in good shape for landing code upstream,
regularly review as independent changes and ease of porting to new each
new stable release.

We deleted the branch pie branch to force explicitly referring to that
code via PQ3B.190801.002.2019.08.25.15. This makes it clear that it is a
snapshot of the project at a certain point in time. In the future, we'll
delete the legacy branches faster to avoid confusion when they're no
longer being used. The tags are how people should refer to the
historical state of the project.